+++ lib/layers/rubocop.rb
@@ -1,0 +1,6 @@
+# frozen_string_literal: true
+
+require 'rubocop'
+require 'rubocop/cop/layers/user_story_reference'
+require 'rubocop/cop/layers/use_case_calls_user_story'
+require 'rubocop/cop/layers/user_story_outside_adapter'